+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Debug logging for Sync
 - Style: barrier color in new measurement modal
 - Style: add measurement icon
+- Style: floating action button color
 
 ## [0.8.145] - 2022-09-15
 ### Changed: